Change texture handling
- Close #6

+++ openvr/src/openvr_render_stage.hpp
@@ -46,7 +46,7 @@ public:
     void reload_shaders() final;
 
     void set_render_target_size(int width, int height);
-    unsigned int get_eye_texture(vr::EVREye vr_eye);
+    uint64_t get_eye_texture(vr::EVREye vr_eye);
 
 private:
     std::string get_plugin_id() const final;
